requirements.txt
openai==1.7.0
python-telegram-bot==20.7
google-generativeai==0.3.2
python-dotenv==1.0.0
pyyaml==6.0